assignment_08/ai_agent_app/requirements.txt
agency-swarm==0.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
datamodel-code-generator==0.26.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==6.3.7
email_validator==2.2.0
fastapi==0.115.5
frozenlist==1.5.0
genson==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.55.3
ordered-set==4.1.0
orjson==3.10.12
packaging==23.2
pathspec==0.12.1
platformdirs==4.3.6
primp==0.8.1
propcache==0.2.0
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.8.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.44.0
yarl==1.18.0
